+++ nowplaying/processes/trackpoll.py
@@ -417,6 +417,12 @@ class TrackPoll:  # pylint: disable=too-many-instance-attributes
             return
 
         # fill in the blanks and make it live
+        logging.debug(
+            "raw input metadata: artist=%s title=%s filename=%s",
+            nextmeta.get("artist"),
+            nextmeta.get("title"),
+            nextmeta.get("filename"),
+        )
         oldmeta = self.currentmeta
         fill_start_time = time.time()
         try: